add validation to adding and updating users

+++ app/Http/Controllers/AdminController.php
@@ -6,6 +6,7 @@ use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\AdminRequest;
+use Illuminate\Validation\Rule;
 
 class AdminController extends Controller
 {
@@ -37,6 +38,23 @@ class AdminController extends Controller
      */
     public function store(Request $request)
     {
+        $request->validate(['username' => [
+            'required'],
+            'password' => [
+                'required',
+                'min:8',
+                'regex:/^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{6,}$/',
+                'confirmed'
+            ],
+            'email' => [
+                'required',
+                'email:filter'
+            ],
+            'privilege' =>[
+                'required',
+                Rule::in(['user', 'manager'])
+            ]
+        ]);
         $user = new User;
 //        request()->all()->password = ;
         $request->merge(['password' => Hash::make(request()->all()['password'])]);
@@ -76,18 +94,38 @@ class AdminController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(AdminRequest $request,$id)
+    public function update(Request $request,$id)
     {
-        $user = User::find($id);
         $requestData = $request->all();
+        $validated = array();
+        $rules = array('username' => [
+                    Rule::unique('users','username')->where(function ($query) {
+                        return $query->where('deleted_at', Null);
+                    })->ignore($id)],
+            'password' => [
+                    'min:8',
+                    'regex:/^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{6,}$/',
+                    'confirmed'
+        ],
+            'email' => [
+                'email:filter'
+            ]);
+        foreach ($rules as $key => $value)
+        {
+            if(isset($requestData[$key])){
+                $validated[$key] = $value;
+            }
+        }
+        $request->validate($validated);
+        $user = User::find($id);
         if(isset($requestData['password'])){
-            $requestData['password'] =Hash::make(request()->all()['password']);
+            $requestData['password'] = Hash::make(request()->all()['password']);
         } else {
             $requestData['password'] = $user->password;
         }
 //        return $requestData;
         User::updateOrCreate(['id'=>$user->id],$requestData);
-        return redirect()->route('all_users', ['users' => User::all()]);
+        return redirect()->route('all_users', ['users' => User::all()])->with('success', 'user updated');
     }
 
     /**
@@ -99,8 +137,13 @@ class AdminController extends Controller
     public function destroy($id)
     {
         //
+
         $user = User::find($id);
+        if($user->privilege == 'manager' && User::where('privilege', 'manager')->count() == 1){
+
+            return redirect()->route('all_users', ['users' => User::all()])->with('error', 'can not delete last manager');
+        }
         $user->delete();
-        return redirect()->route('all_users', ['users' => User::all()]);
+        return redirect()->route('all_users', ['users' => User::all()])->with('success', 'user deleted');
     }
 }
